requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas>=2.2.0
numpy==1.24.3
Pillow==10.1.0
python-dotenv==1.0.0
apscheduler==3.10.4
openai==1.3.7
praw==7.7.1
yfinance==0.2.18
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1 